FedMLGCPRetailUseCase/RetailTestPredictor/requirements.txt
fastapi
uvicorn==0.17.6
joblib~=1.0
numpy~=1.20
scikit-learn==0.23.2
google-cloud-storage>=1.26.0,<2.0.0dev
google-cloud-aiplatform
pandas
lightgbm
fedml_gcp